requirements.txt
boto3>=1.34.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
Pillow>=11.0.0,<12.0.0
